Introduce dnf.py for dnf related logic
We put this in a new "installer" submodule as well.
