+++ mkosi/distributions/opensuse.py
@@ -8,7 +8,7 @@ from collections.abc import Sequence
 
 from mkosi.architecture import Architecture
 from mkosi.distributions import DistributionInstaller
-from mkosi.distributions.fedora import Repo, fixup_rpmdb_location, invoke_dnf, setup_dnf
+from mkosi.installer.dnf import Repo, fixup_rpmdb_location, invoke_dnf, setup_dnf
 from mkosi.log import die
 from mkosi.run import bwrap
 from mkosi.state import MkosiState
