+++ mkosi/installer/__init__.py
@@ -1,0 +1,1 @@
+# SPDX-License-Identifier: LGPL-2.1+
